V19.09.22.00 Debuggability matters more.

+++ include/object/shape/shape.h
@@ -52,7 +52,7 @@ public:
     void addBSSRDF(BSSRDF *b);
     BSSRDF* getBSSRDF();
 
-    virtual void simplify(const float &threshold, const int &pointsLeft) = 0;
+    virtual void simplify(const float &threshold, const float &pointsLeft) = 0;
 
 };
